pyproject.toml
[project]
# This is the name of your project. The first time you publish this
# package, this name will be registered for you. It will determine how
# users can install this project, e.g.:
#
# $ pip install sampleproject
#
# And where it will live on PyPI: https://pypi.org/project/sampleproject/
#
# There are some restrictions on what makes a valid project name
# specification here:
# https://packaging.python.org/specifications/core-metadata/#name
name = "classifier"  # Required

# Versions should comply with PEP 440:
# https://www.python.org/dev/peps/pep-0440/
#
# For a discussion on single-sourcing the version, see
# https://packaging.python.org/guides/single-sourcing-package-version/
version = "0.1.20"  # Required

# This is a one-line description or tagline of what your project does. This
# corresponds to the "Summary" metadata field:
# https://packaging.python.org/specifications/core-metadata/#summary
description = "Entity Classifier"  # Optional

# This is an optional longer description of your project that represents
# the body of text which users will see when they visit PyPI.
#
# Often, this is the same as your README, so you can just read it in from
# that file directly (as we have already done above)
#
# This field corresponds to the "Description" metadata field:
# https://packaging.python.org/specifications/core-metadata/#description-optional
# readme = "README.md" # Optional - commented out as file doesn't exist

# Specify which Python versions you support. In contrast to the
# 'Programming Language' classifiers above, 'pip install' will check this
# and refuse to install the project if the version does not match. See
# https://packaging.python.org/guides/distributing-packages-using-setuptools/#python-requires
requires-python = ">=3.9"

# This is either text indicating the license for the distribution, or a file
# that contains the license
# https://packaging.python.org/en/latest/specifications/core-metadata/#license
# Using text format since "Proprietary" is not a valid SPDX identifier
license = {text = "Proprietary"}

# This field adds keywords for your project which will appear on the
# project page. What does your project relate to?
#
# Note that this is a list of additional keywords, separated
# by commas, to be used to assist searching for the distribution in a
# larger catalog.
keywords = ["langchain", "ai", "rag"]  # Optional



# Classifiers help users find your project by categorizing it.
#
# For a list of valid classifiers, see https://pypi.org/classifiers/
classifiers = [  # Optional
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Developers",

  "Topic :: Software Development :: Build Tools",

  # License is now specified using SPDX expression in project.license above
  # "License :: Other/Proprietary License",  # Deprecated - removed

  "Operating System :: OS Independent",

  # Specify the Python versions you support here. In particular, ensure
  # that you indicate you support Python 3. These classifiers are *not*
  # checked by "pip install". See instead "python_requires" below.
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  
]

# This field lists other packages that your project depends on to run.
# Any package you put here will be installed by pip when your project is
# installed, so they must be valid existing projects.
#
# For an analysis of this field vs pip's requirements files see:
# https://packaging.python.org/discussions/install-requires-vs-requirements/
dependencies = [ 
  "fastapi==0.109.1",
  "uvicorn==0.25.0",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "presidio-analyzer==2.2.358",
  "presidio-anonymizer==2.2.358",
  "torch>=2.1.2",
  "transformers==4.51.3",
  "jinja2>=3.1.3",
  "tqdm",
  "python-dateutil==2.9.0",
  "openai== 1.99.9",
  "litellm==1.74.3",
  "json-repair==0.44.1",
  "boto3==1.35.58",
  "python-multipart==0.0.17",
  "python-stdnum==1.20",
  "luhnchecker==0.0.12",
  "langchain==1.0.5",
  "langchain-core==1.0.4",
  "backoff>=2.2.1"  # Added for exponential backoff retry logic
]

# List additional groups of dependencies here (e.g. development
# dependencies). Users will be able to install these using the "extras"
# syntax, for example:
#
#   $ pip install sampleproject[dev]
#   For zsh users, use following syntax:
#   $ pip install -e 'sampleproject[dev]'
# or $ pip install -e '.[dev]'
#
# Similar to `dependencies` above, these must be valid existing
# projects.
[project.optional-dependencies] # Optional
dev = ["check-manifest"]
test = ["ruff"]

# List URLs that are relevant to your project
#
# This field corresponds to the "Project-URL" and "Home-Page" metadata fields:
# https://packaging.python.org/specifications/core-metadata/#project-url-multiple-use
# https://packaging.python.org/specifications/core-metadata/#home-page-optional
#
# Examples listed include a pattern for specifying where the package tracks
# issues, where the source is hosted, where to say thanks to the package
# maintainers, and where to support the project financially. The key is
# what's used to render the link text on PyPI.


# The following would provide a command line executable called `sample`
# which executes the function `main` from this package when invoked.
[project.scripts]  # Optional
classifier = "classifier.app.daemon:start"

# This is configuration specific to the `setuptools` build backend.
# If you are using a different build backend, you will need to change this.
[tool.setuptools]
# Use find directive to automatically discover all subpackages
# This prevents setuptools from including unwanted directories like evals, myenv, deploy
packages = {find = {exclude = ["evals*", "myenv*", "deploy*", "build*", "dist*", "*.egg-info*"]}}
include-package-data = false

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"


[tool.mypy]
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
